requirements.txt
pydantic-settings==2.12.0
prisma==0.15.0
pynest-api==0.4.0
python-jose[cryptography]==3.3.0
fastapi_swagger_dark==0.0.9